add refresh session

+++ src/services/auth.js
@@ -17,6 +17,7 @@ export const registerUser = async (payload) => {
 
 export const loginUser = async (email, password) => {
   const maybeUser = await UsersCollection.findOne({ email });
+
   if (maybeUser === null) {
     throw createHttpError(404, 'User not found');
   }
@@ -27,18 +28,40 @@ export const loginUser = async (email, password) => {
 
   await Session.deleteOne({ userId: maybeUser._id });
 
-  const accessToken = crypto.randomBytes(30).toString('base64');
-  const refreshToken = crypto.randomBytes(30).toString('base64');
-
   return Session.create({
-    userId: maybeUser,
-    accessToken,
-    refreshToken,
+    userId: maybeUser._id,
+    accessToken: crypto.randomBytes(30).toString('base64'),
+    refreshToken: crypto.randomBytes(30).toString('base64'),
     accessTokenValidUntil: new Date(Date.now() + ACCESS_TOKEN_TTL),
     refreshTokenValidUntil: new Date(Date.now() + REFRESH_TOKEN_TTL),
   });
 };
 
-export const logOutUser = (sessionId) => {
-  return Session.deleteOne({ _id: sessionId });
+export const logOutUser = async (sessionId) => {
+  await Session.deleteOne({ _id: sessionId });
+};
+
+export const refreshUserSession = async (sessionId, refreshToken) => {
+  const session = await Session.findOne({
+    _id: sessionId,
+    refreshToken,
+  });
+
+  if (session === null) {
+    throw createHttpError(401, 'Session not found');
+  }
+
+  if (new Date() > new Date(session.refreshTokenValidUntil)) {
+    throw createHttpError(401, 'Refresh token is expired');
+  }
+
+  await Session.deleteOne({ _id: sessionId });
+
+  return Session.create({
+    userId: session.userId,
+    accessToken: crypto.randomBytes(30).toString('base64'),
+    refreshToken: crypto.randomBytes(30).toString('base64'),
+    accessTokenValidUntil: new Date(Date.now() + ACCESS_TOKEN_TTL),
+    refreshTokenValidUntil: new Date(Date.now() + REFRESH_TOKEN_TTL),
+  });
 };
